+++ app/Providers/AppServiceProvider.php
@@ -4,6 +4,9 @@ namespace App\Providers;
 
 use App\Services\Hashid;
 use App\Services\Shaarli\Shaarli;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\View;
 use Illuminate\Support\ServiceProvider;
 
@@ -46,6 +49,10 @@ class AppServiceProvider extends ServiceProvider
             } else {
                 $view->with('lang', []);
             }
+        });
+
+        Validator::extend('current_password', function ($attribute, $value, $parameters, $validator) {
+            return Hash::check($value, Auth::user()->getAuthPassword());
         });
     }
 }