+++ resources/views/manage/logins.blade.php
@@ -61,6 +61,28 @@
                 @endif
             </div>
         </div>
+
+        <div class="card mt-4">
+            <div class="card-header">{{ __('Logout from other devices') }}</div>
+
+            <div class="card-body">
+                <p>{{ __("Type your current password to logout from all other sessions.") }}</p>
+
+                <form action="{{ route('manage.logins.logout') }}" method="POST">
+                    @csrf
+
+                    <div class="form-group">
+                        <label for="password">{{ __('Password') }}</label>
+                        <input type="password" name="password" id="password" class="form-control @error('password') is-invalid @enderror" required />
+                        @error('password')
+                        <span class="invalid-feedback" role="alert">{{ $message }}</span>
+                        @enderror
+                    </div>
+
+                    <button type="submit" class="btn btn-outline-danger">{{ __('Confirm') }}</button>
+                </form>
+            </div>
+        </div>
     </div>
 </div>
 @endsection