+++ routes/web.php
@@ -39,5 +39,6 @@ Route::get('manage/export', 'ManageController@exportForm')->name('manage.export'
 Route::post('manage/export', 'ManageController@export');
 Route::get('manage/tags', 'ManageController@tags')->name('manage.tags');
 Route::get('manage/logins', 'ManageController@logins')->name('manage.logins');
+Route::post('manage/logins/logout', 'ManageController@logoutDevices')->name('manage.logins.logout');
 Route::get('manage/settings', 'ManageController@settingsForm')->name('manage.settings');
 Route::post('manage/settings', 'ManageController@settingsStore');